Remove DeckGLOverlay in geojson example
