+++ website/src/main.js
@@ -6,8 +6,11 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 
-import AppState from './reducers';
-import Routes from './routes';
+/* global window */
+window.website = true;
+// Use require to ensure `window.website` flag is set before demos are imported.
+const AppState = require('./reducers').default;
+const Routes = require('./routes').default;
 
 const el = document.createElement('div');
 document.body.appendChild(el);
